refactor(rust/sedona-raster-gdal): share geometry-mask machinery between RS_Clip and RS_ZonalStats
Extract the pixel-window addressing and geometry rasterization that RS_Clip
and RS_ZonalStats each carried into a shared mask module: one PixelWindow,
one envelope_window, one rasterize_geometry_mask. RS_Clip now builds its
geotransform from raster.transform() instead of an inline coefficient array.
Behavior is unchanged; the deliberately-different per-pixel consumers
(apply_mask_to_band writes nodata, collect_masked_values reads f64) stay
separate.
